Implemented the 'add' method and inserted some rows into my 'client' table.

+++ magasin/src/products/IProduitDao.java
@@ -4,8 +4,8 @@ import java.util.*;
 public interface IProduitDao {
 	//abstract methods
 	public void add(Produit p);
-	public void delete(long id);
-	public Produit getOne(long id);
+	public void delete(int id);
+	public Produit getOne(int id);
 	public List<Produit> getAll();
 	public List<Produit> getAll(String des);
